+++ src/strings.rs
@@ -558,6 +558,16 @@ pub mod commands {
 			CMD_GROUP_LOG,
 		)
 	}
+	pub fn copy_path(key_config: &SharedKeyConfig) -> CommandText {
+		CommandText::new(
+			format!(
+				"Copy Path [{}]",
+				key_config.get_hint(key_config.keys.copy),
+			),
+			"copy selected file path to clipboard",
+			CMD_GROUP_LOG,
+		)
+	}
 	pub fn push_tags(key_config: &SharedKeyConfig) -> CommandText {
 		CommandText::new(
 			format!(